+++ plot3d.py
@@ -10,9 +10,10 @@ import prob
 ############################################
 # set default data type to double
 ############################################
-torch.set_default_dtype(torch.float64)
-torch.set_default_tensor_type(torch.DoubleTensor)
-
+# torch.set_default_dtype(torch.float64)
+# torch.set_default_tensor_type(torch.DoubleTensor)
+torch.set_default_dtype(torch.float32)
+torch.set_default_tensor_type(torch.FloatTensor)
 
 ###########################################
 # plot function for 3-d systems
